fix(luasnip) adapt to newest update
